backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
drf-yasg==1.21.7
django-extensions==3.2.3
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
whitenoise==6.6.0
Pillow==10.1.0
python-decouple==3.8
gunicorn==21.2.0
stripe==7.6.0 